+++ common/src/main/java/net/kroia/stockmarket/networking/packet/client_sender/update/entity/UpdateStockMarketBlockEntityPacket.java
@@ -11,7 +11,7 @@ public class UpdateStockMarketBlockEntityPacket extends StockMarketNetworkPacket
     private BlockPos pos;
     private TradingPair tradingPair;
     private int amount;
-    private int price;
+    private float price;
 
 
 
@@ -40,7 +40,7 @@ public class UpdateStockMarketBlockEntityPacket extends StockMarketNetworkPacket
         return amount;
     }
 
-    public int getPrice() {
+    public float getPrice() {
         return price;
     }
 
@@ -54,7 +54,7 @@ public class UpdateStockMarketBlockEntityPacket extends StockMarketNetworkPacket
         buf.writeBlockPos(pos);
         tradingPair.encode(buf);
         buf.writeInt(amount);
-        buf.writeInt(price);
+        buf.writeFloat(price);
     }
 
     @Override
@@ -67,7 +67,7 @@ public class UpdateStockMarketBlockEntityPacket extends StockMarketNetworkPacket
         }
         this.tradingPair.decode(buf);
         this.amount = buf.readInt();
-        this.price = buf.readInt();
+        this.price = buf.readFloat();
     }
 
     @Override
